requirements.txt
# Core data processing
pandas>=2.3.0
numpy>=2.2.0

# Machine Learning and Embeddings
sentence-transformers>=5.1.0
umap-learn>=0.5.9
scikit-learn>=1.7.0

# Progress bars
tqdm>=4.67.0

# HTTP requests (for model downloads)
requests>=2.32.0